requirements.txt
streamlit
pandas
numpy
scikit-learn
mysql-connector-python
joblib
matplotlib
openpyxl